fix(runs): let the missing-signal gate stay silent past the bound
The bound was checked before the provider signal, so a run whose stream
holds no reap.provider_error message emitted reap.provider_retry_exhausted
once its lineage was spent. The design doc says that case returns without an
event, because no classifier ever ran and a verdict would describe a
decision nobody made. The signal gate now comes first, which also reads the
events already in hand before the bound reads chain rows, and the doc says
so.

Two tests: the provider, model and cap still ride the last retry the bound
allows, and a run with no signal writes no verdict even when exhausted.

The suite split rather than raise a budget. provider-retry.test.ts sat one
line under the 500-line ceiling on main, so the bound cases moved to
provider-retry.bound.test.ts and the fixture the two share moved to
provider-retry.test-fixture.ts.
